requirements.txt
langchain==0.3.x
langchain-openai==0.2.x
langchain-community==0.3.x
langchain-core==0.3.x
openai>=1.40.0
faiss-cpu==1.8.0
chromadb==0.5.x
sentence-transformers==3.x
pypdf2==3.x
pdfplumber==0.11.x
fastapi==0.115.x
uvicorn[standard]==0.32.x
streamlit==1.40.x
mlflow==2.18.x
ragas==0.2.x
deepeval==1.x
python-dotenv==1.0.x
tiktoken==0.8.x
datasets==3.x